// src/games/dto/basic-game/basic-participant.dto.ts
export type BasicParticipantDto = {
  assists: number;
  championId: number;
  championName: string;
  champLevel: number;
  deaths: number;
  kills: number;
  placement: number;
  playerSubTeamId: number;
  position: string;
  riotIdGameName: string;
  riotIdTagline: string;
  summonerName: string;
  teamId: number;
};